requirements.txt
cohere
python-dotenv
requests
atproto
